+++ glean/src/platform/browser/web/index.ts
@@ -4,7 +4,7 @@
 
 import type Platform from "../../index.js";
 
-import uploader from "../fetch_uploader.js";
+import uploader from "../sendbeacon_fallback_uploader.js";
 import info from "./platform_info.js";
 import Storage from "./storage.js";