+++ CHANGELOG.md
@@ -1,12 +1,20 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean.js/compare/v4.0.0...main)
+[Full changelog](https://github.com/mozilla/glean.js/compare/v4.1.0-pre.0...main)
+
+# v4.1.0-pre.0 (2024-03-05)
+
+[Full changelog](https://github.com/mozilla/glean.js/compare/v4.0.0...v4.1.0-pre.0)
+
+* [#1866](https://github.com/mozilla/glean.js/pull/1866): Added a new uploader that falls back to `fetch` if `navigator.sendBeacon` fails. 
+* [#1876](https://github.com/mozilla/glean.js/pull/1876): **BREAKING CHANGE**: `navigator.sendBeacon` with fallback to `fetch` (see #1866) is now the default uploader. This can be changed manually.
+* [#1850](https://github.com/mozilla/glean.js/pull/1850): Automatically record basic session information (`session_id` & `session_count`) for web properties.
 
 # v4.0.0 (2024-01-24)
 
 [Full changelog](https://github.com/mozilla/glean.js/compare/v4.0.0-pre.3...v4.0.0)
 
-* This is the official release based on the v4.x.x-pre.x releases.
+* This is the official release based on the v4.0.0-pre.x releases.
 
 # v4.0.0-pre.3 (2023-12-22)